requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# Deep Learning (optional - comment out if not needed)
# tensorflow>=2.13.0
# torch>=2.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Feature Store (optional - comment out if not using)
# hopsworks>=3.0.0

# Model Registry (optional - comment out if not using)
# mlflow>=2.0.0

# Jupyter
# jupyter>=1.0.0
# ipykernel>=6.25.0

# Model Explainability
shap>=0.42.0

# Web Framework (for future dashboard)
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
